+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 All notable changes to the "pomocode" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+## [0.1.0] - 2025-11-30
+
+### Added
+
+* Status bar tomato icon for starting sessions.
+* Quick Pomocode Session (default 25 minutes).
+* Quick break session (default 5 minutes)
+* Status bar timer with pause/resume functionality.
+* Automatic short break prompt after each session.
+* Time display formatting (HH:MM:SS).
+* Notification messages for session completion and break end.